wip(theme-builder): abstract out theme state into provider
