+++ website/src/pages/themes/_config/chart.tsx
@@ -12,425 +12,17 @@ import {
   VictoryLegend,
   VictoryLine,
   VictoryPie,
-  VictoryPolarAxis,
   VictoryScatter,
   VictoryVoronoi,
 } from "victory";
-import { colorScaleOptions } from "./_components/color-scale-options";
+import {
+  getBaseLabelsConfig,
+  getBaseStrokeConfig,
+  getNestedColorScaleConfig,
+  StrokeProps,
+} from "../_utils";
 
-type ThemeBuilderFieldConfig =
-  | {
-      type: "section" | "colorScale";
-      label: string;
-      fields?: ThemeBuilderFieldConfig[];
-    }
-  | {
-      type: "slider" | "select" | "colorPicker";
-      label: string;
-      path: string | string[];
-      min?: number;
-      max?: number;
-      step?: number;
-      unit?: string;
-      options?: { label: string; value: string }[];
-    };
-
-enum StrokeProps {
-  STROKE = "Stroke",
-  STROKE_WIDTH = "Stroke Width",
-  STROKE_DASH_ARRAY = "Stroke Dash Array",
-  STROKE_LINE_CAP = "Stroke Line Cap",
-  STROKE_LINE_JOIN = "Stroke Line Join",
-}
-
-type ThemeBuilderOptionsConfig = {
-  title: string;
-  hasVictoryChart?: boolean;
-  content?: (props: any) => React.ReactNode;
-  fields: ThemeBuilderFieldConfig[];
-}[];
-
-const getPath = (basePath: string | string[], key: string) => {
-  if (Array.isArray(basePath)) {
-    return basePath.map((p) => `${p}.${key}`);
-  }
-  return `${basePath}.${key}`;
-};
-
-const getNestedColorScaleConfig = (
-  basePath: string | string[],
-): ThemeBuilderFieldConfig[] => [
-  {
-    type: "select",
-    label: "Color Scale",
-    options: colorScaleOptions,
-    path: getPath(basePath, "colorScale"),
-  },
-];
-
-const getBaseStrokeConfig = (
-  basePath: string | string[],
-  includedStrokeProps: StrokeProps[] = [],
-): ThemeBuilderFieldConfig[] => {
-  const config = [
-    {
-      type: "colorPicker",
-      label: StrokeProps.STROKE,
-      path: getPath(basePath, "stroke"),
-    },
-    {
-      type: "slider",
-      label: StrokeProps.STROKE_WIDTH,
-      min: 0,
-      max: 5,
-      unit: "px",
-      path: getPath(basePath, "strokeWidth"),
-    },
-    {
-      type: "slider",
-      label: StrokeProps.STROKE_DASH_ARRAY,
-      min: 0,
-      max: 10,
-      path: getPath(basePath, "strokeDasharray"),
-    },
-    {
-      type: "select",
-      label: StrokeProps.STROKE_LINE_CAP,
-      options: [
-        { label: "Round", value: "round" },
-        { label: "Square", value: "square" },
-        { label: "Butt", value: "butt" },
-      ],
-      path: getPath(basePath, "strokeLinecap"),
-    },
-    {
-      type: "select",
-      label: StrokeProps.STROKE_LINE_JOIN,
-      options: [
-        { label: "Round", value: "round" },
-        { label: "Bevel", value: "bevel" },
-        { label: "Miter", value: "miter" },
-      ],
-      path: getPath(basePath, "strokeLinejoin"),
-    },
-  ] as ThemeBuilderFieldConfig[];
-  return includedStrokeProps.length
-    ? config.filter((field) =>
-        includedStrokeProps.includes(field.label as StrokeProps),
-      )
-    : config;
-};
-
-const getBaseLabelsConfig = (
-  basePath: string | string[],
-): ThemeBuilderFieldConfig[] => [
-  {
-    type: "slider",
-    label: "Font Size",
-    min: 10,
-    max: 24,
-    unit: "px",
-    path: getPath(basePath, "fontSize"),
-  },
-  {
-    type: "slider",
-    label: "Text Padding",
-    min: 0,
-    max: 50,
-    unit: "px",
-    path: getPath(basePath, "padding"),
-  },
-  {
-    type: "colorPicker",
-    label: "Font Color",
-    path: getPath(basePath, "fill"),
-  },
-];
-
-export const paletteOptionsConfig = {
-  title: "Color Palette",
-  description: "Choose custom colors for each color scale.",
-  fields: colorScaleOptions.map((option) => ({
-    type: "colorScale",
-    colorScaleType: option.value,
-    label: option.label,
-  })),
-};
-
-export const globalOptionsConfig = {
-  title: "Global Options",
-  fields: [
-    {
-      type: "section",
-      label: "Chart Size",
-      description: "Set the width, height, and padding of the chart.",
-      fields: [
-        {
-          type: "slider",
-          label: "Width",
-          min: 0,
-          max: 500,
-          unit: "px",
-          path: [
-            "chart.width",
-            "axis.width",
-            "area.width",
-            "bar.width",
-            "boxplot.width",
-            "candlestick.width",
-            "errorbar.width",
-            "group.width",
-            "histogram.width",
-            "line.width",
-            "pie.width",
-            "scatter.width",
-            "stack.width",
-            "voronoi.width",
-          ],
-        },
-        {
-          type: "slider",
-          label: "Height",
-          min: 0,
-          max: 500,
-          unit: "px",
-          path: [
-            "chart.height",
-            "axis.height",
-            "area.height",
-            "bar.height",
-            "boxplot.height",
-            "candlestick.height",
-            "errorbar.height",
-            "group.height",
-            "histogram.height",
-            "line.height",
-            "pie.height",
-            "scatter.height",
-            "stack.height",
-            "voronoi.height",
-          ],
-        },
-        {
-          type: "slider",
-          label: "Padding",
-          min: 0,
-          max: 100,
-          unit: "px",
-          path: [
-            "chart.padding",
-            "axis.padding",
-            "area.padding",
-            "bar.padding",
-            "boxplot.padding",
-            "candlestick.padding",
-            "errorbar.padding",
-            "group.padding",
-            "histogram.padding",
-            "line.padding",
-            "pie.padding",
-            "scatter.padding",
-            "stack.padding",
-            "voronoi.padding",
-          ],
-        },
-      ],
-    },
-    {
-      type: "section",
-      label: "Labels",
-      fields: getBaseLabelsConfig([
-        "axis.style.axisLabel",
-        "polarAxis.style.tickLabels",
-        "polarDependentAxis.style.tickLabels",
-        "tooltip.style",
-        "area.style.labels",
-        "bar.style.labels",
-        "candlestick.style.labels",
-        "errorbar.style.labels",
-        "histogram.style.labels",
-        "legend.style.labels",
-        "line.style.labels",
-        "pie.style.labels",
-        "scatter.style.labels",
-        "voronoi.style.labels",
-        "boxplot.style.maxLabels",
-        "boxplot.style.medianLabels",
-        "boxplot.style.minLabels",
-        "boxplot.style.q1Labels",
-        "boxplot.style.q3Labels",
-      ]),
-    },
-    {
-      type: "section",
-      label: "Data",
-      fields: getBaseStrokeConfig([
-        "area.style.data",
-        "bar.style.data",
-        "candlestick.style.data",
-        "errorbar.style.data",
-        "histogram.style.data",
-        "line.style.data",
-        "pie.style.data",
-        "scatter.style.data",
-        "voronoi.style.data",
-      ]),
-    },
-  ],
-};
-
-const optionsConfig: ThemeBuilderOptionsConfig = [
-  {
-    title: "Color Palette",
-    fields: [
-      {
-        type: "colorScale",
-        label: "Color Scale",
-      },
-    ],
-  },
-  {
-    type: "section",
-    title: "Axis Options",
-    fields: [
-      {
-        type: "section",
-        label: "General",
-        fields: getBaseStrokeConfig("axis.style.axis", [
-          StrokeProps.STROKE,
-          StrokeProps.STROKE_WIDTH,
-          StrokeProps.STROKE_LINE_CAP,
-          StrokeProps.STROKE_LINE_JOIN,
-        ]),
-      },
-      {
-        type: "section",
-        label: "Grid",
-        fields: getBaseStrokeConfig("axis.style.grid"),
-      },
-      {
-        type: "section",
-        label: "Ticks",
-        fields: [
-          {
-            type: "slider",
-            label: "Size",
-            min: 0,
-            max: 50,
-            unit: "px",
-            path: "axis.style.ticks.size",
-          },
-          ...getBaseStrokeConfig("axis.style.ticks", [
-            StrokeProps.STROKE,
-            StrokeProps.STROKE_WIDTH,
-          ]),
-        ],
-      },
-      {
-        type: "section",
-        label: "Labels",
-        fields: getBaseLabelsConfig("axis.style.axisLabel"),
-      },
-    ],
-  },
-  {
-    type: "section",
-    title: "Polar Axis",
-    content: (props) => [
-      <VictoryPolarAxis {...props} key="polar-axis" standalone={false} />,
-      <VictoryPolarAxis
-        {...props}
-        key="polar-axis-dependent"
-        dependentAxis
-        domain={[0, 10]}
-        standalone={false}
-      />,
-    ],
-    fields: [
-      {
-        type: "section",
-        label: "General",
-        fields: getBaseStrokeConfig("polarAxis.style.axis", [
-          StrokeProps.STROKE,
-          StrokeProps.STROKE_WIDTH,
-        ]),
-      },
-      {
-        type: "section",
-        label: "Grid",
-        fields: getBaseStrokeConfig("polarAxis.style.grid"),
-      },
-      {
-        type: "section",
-        label: "Ticks",
-        fields: [
-          {
-            type: "slider",
-            label: "Size",
-            min: 0,
-            max: 50,
-            unit: "px",
-            path: "polarAxis.style.ticks.size",
-          },
-          ...getBaseStrokeConfig("polarAxis.style.ticks", [
-            StrokeProps.STROKE,
-            StrokeProps.STROKE_WIDTH,
-            StrokeProps.STROKE_LINE_CAP,
-            StrokeProps.STROKE_LINE_JOIN,
-          ]),
-        ],
-      },
-      {
-        type: "section",
-        label: "Labels",
-        fields: getBaseLabelsConfig("polarAxis.style.tickLabels"),
-      },
-    ],
-  },
-  {
-    type: "section",
-    title: "Polar Dependent Axis",
-    fields: [
-      {
-        type: "section",
-        label: "General",
-        fields: getBaseStrokeConfig("polarDependentAxis.style.axis", [
-          StrokeProps.STROKE,
-          StrokeProps.STROKE_WIDTH,
-        ]),
-      },
-      {
-        type: "section",
-        label: "Grid",
-        fields: getBaseStrokeConfig("polarDependentAxis.style.grid"),
-      },
-      {
-        type: "section",
-        label: "Ticks",
-        fields: [
-          {
-            type: "slider",
-            label: "Size",
-            min: 0,
-            max: 50,
-            unit: "px",
-            path: "polarDependentAxis.style.ticks.size",
-          },
-          ...getBaseStrokeConfig("polarDependentAxis.style.ticks", [
-            StrokeProps.STROKE,
-            StrokeProps.STROKE_WIDTH,
-            StrokeProps.STROKE_LINE_CAP,
-            StrokeProps.STROKE_LINE_JOIN,
-          ]),
-        ],
-      },
-      {
-        type: "section",
-        label: "Labels",
-        fields: getBaseLabelsConfig("polarDependentAxis.style.tickLabels"),
-      },
-    ],
-  },
+export default [
   {
     type: "section",
     title: "Area Chart",
@@ -1223,5 +815,3 @@ const optionsConfig: ThemeBuilderOptionsConfig = [
     ],
   },
 ];
-
-export default optionsConfig;
